# specs/openapi/adscenter.yaml
openapi: 3.0.3
info:
  title: Adscenter Service
  version: 0.1.0
  license:
    name: Elastic-Internal
servers:
  - url: /
paths:
...
  /api/v1/adscenter/bulk-actions:
    get:
      operationId: listBulkActions
      summary: List recent bulk action operations for current user
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/BulkActionOperation'
        '401': { description: Unauthorized }
    post:
      operationId: submitBulkActions
      summary: Submit bulk actions
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/BulkActionPlan'
                - $ref: '#/components/schemas/OpportunityComboPlan'
      responses:
        '200': { description: OK (validateOnly) }
        '202': { description: Accepted (enqueued) }
        '400': { description: Bad Request }
        '401': { description: Unauthorized }
...
  /api/v1/adscenter/bulk-actions/{id}/rollback:
    post:
      operationId: rollbackBulkAction
      summary: Mark operation as rolled back and append a rollback snapshot (stub)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: Rolled back }
        '401': { description: Unauthorized }
        '404': { description: Not Found }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    BulkActionOperation:
      type: object
      properties:
        operationId: { type: string }
        status: { type: string, enum: [queued, running, completed, failed] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        summary:
          type: object
          properties:
            actions: { type: integer }
            estimatedAffected: { type: integer }
      required: [operationId, status]
    BulkActionPlan:
      type: object
      properties:
        validateOnly:
          type: boolean
        actions:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
                enum: [ADJUST_CPC, ADJUST_BUDGET, ADD_NEGATIVE_KEYWORDS, REMOVE_NEGATIVE_KEYWORDS, PAUSE_KEYWORDS, ENABLE_KEYWORDS, SET_AD_SCHEDULES, SET_TARGET_CPA, SET_TARGET_ROAS]
              filter:
                type: object
                additionalProperties: true
              params:
                oneOf:
                  - $ref: '#/components/schemas/AdjustCpcParams'
                  - $ref: '#/components/schemas/AdjustBudgetParams'
                  - $ref: '#/components/schemas/NegativeKeywordsParams'
                  - $ref: '#/components/schemas/KeywordStatusParams'
                  - $ref: '#/components/schemas/AdSchedulesParams'
                  - $ref: '#/components/schemas/BiddingParams'
          minItems: 1
      required: [actions]
    AdjustCpcParams:
      type: object
      properties:
        keyword: { type: string, description: Target keyword text }
        percent:
          type: number
          description: "Increase/decrease percent (e.g., 10 for +10%)"
        cpcValue: { type: number, description: Absolute CPC value to set (optional alternative to percent) }
        reason: { type: string }
        seedDomain: { type: string }
        country: { type: string }
    AdjustBudgetParams:
      type: object
      properties:
        dailyBudget: { type: number, description: New daily budget value }
        percent:
          type: number
          description: "Percent change; if present, dailyBudget may be omitted"
    NegativeKeywordsParams:
      type: object
      description: Parameters for adding/removing negative keywords at ad group level
      properties:
        adGroupResourceNames:
          type: array
          items: { type: string }
        keywords:
          type: array
          items: { type: string }
        matchType:
          type: string
          enum: [EXACT, PHRASE, BROAD]
      required: [adGroupResourceNames, keywords]
    KeywordStatusParams:
      type: object
      description: Parameters for pausing/enabling keyword criteria
      properties:
        criterionResourceNames:
          type: array
          items: { type: string }
      required: [criterionResourceNames]
    AdSchedulesParams:
      type: object
      description: Replace campaign ad schedules with provided list
      properties:
        campaignResourceNames:
          type: array
          items: { type: string }
        schedules:
          type: array
          items:
            type: object
            properties:
              dayOfWeek: { type: string, enum: [MONDAY,TUESDAY,WEDNESDAY,THURSDAY,FRIDAY,SATURDAY,SUNDAY] }
              startHour: { type: integer, minimum: 0, maximum: 23 }
              startMinute: { type: string, enum: [ZERO,FIFTEEN,THIRTY,FORTY_FIVE] }
              endHour: { type: integer, minimum: 0, maximum: 23 }
              endMinute: { type: string, enum: [ZERO,FIFTEEN,THIRTY,FORTY_FIVE] }
      required: [campaignResourceNames, schedules]
    BiddingParams:
      type: object
      description: Set bidding targets for campaigns (tCPA/tROAS)
      properties:
        campaignResourceNames:
          type: array
          items: { type: string }
        targetCpaMicros: { type: integer, description: Absolute target CPA in micros }
        targetRoas: { type: number, description: Absolute target ROAS value }
        percent: { type: number, description: Relative adjustment percent }
      required: [campaignResourceNames]
    OpportunityComboPlan:
      type: object
      description: Minimal combo plan built from an Opportunity (keywords/domains) without enumerating low-level actions.
      properties:
        validateOnly: { type: boolean }
        seedDomain: { type: string }
        country: { type: string }
        plan:
          type: object
          properties:
            keywords:
              type: array
              items:
                type: object
                properties:
                  keyword: { type: string }
                  score: { type: number }
                  reason: { type: string }
            domains:
              type: array
              items:
                type: object
                properties:
                  domain: { type: string }
                  score: { type: number }
...